requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.3
SQLAlchemy==2.0.30
pdfminer.six==20221105
scikit-learn==1.5.0
selenium==4.21.0
webdriver-manager==4.0.1
google-generativeai==0.5.4
python-dotenv==1.0.1
